Initial attempt at implementing referenceTarget, plus importing the relevant WPT tests.
https://bugs.webkit.org/show_bug.cgi?id=285634

Reviewed by NOBODY (OOPS!).

This change adds support for a referenceTarget property on ShadowRoot and ShadowRootInit, and a shadowrootreferencetarget property on Templates.
It also adds support code to ensure that the referenceTarget is respected when computing values for internal use, such as creating the accessibility tree.

These additions are behind a compile-time flag and a feature flag.